slableaktracker: Track down sources of slab leaks
The kernel recently allowed for printing the name of a slab cache on
the static tracepoint, `kmem:kmem_cache_alloc`. Now that both the alloc
and corresponding `kmem:kmem_cache_free` static tracepoints provide the
name of the slab cache, leverage these to track down slab leaks in a
slab cache.

Signed-off-by: Charles Haithcock <chaithco@redhat.com>